Trekk ut varsel om motregning ved simulering-logikk til en egen service utenfor SimuleringService
